[feat] PresentaionAssembly, DomainAssembly를 하위 assembly들로 나눔
