--- Planus/Presentation/Application/DI/TmpInjectorPool/PresentationAssembly.swift
@@ -1,552 +1,0 @@
-//
-//  PresentationAssembly.swift
-//  Planus
-//
-//  Created by Sangmin Lee on 3/9/24.
-//
-
-import Foundation
-import Swinject
-
-class PresentationAssembly: Assembly {
-    func assemble(container: Swinject.Container) {
-        assembleMyPageMain(container: container)
-        assembleMyPageReadableViewModel(container: container)
-        assembleMyPageEnquire(container: container)
-        assembleMyPageEdit(container: container)
-        
-        assembleHomeCalendar(container: container)
-        assembleDailyCalendar(container: container)
-        assembleTodoDetail(container: container)
-        
-        assembleSignIn(container: container)
-        assembleRedirectionalWeb(container: container)
-        
-        assembleSearchHome(container: container)
-        assembleSearchResult(container: container)
-        
-        assembleGroupCreate(container: container)
-        assembleGroupCreateLoading(container: container)
-        
-        assembleGroupIntroduce(container: container)
-        
-        assembleNotification(container: container)
-        
-        assembleMyGroupList(container: container)
-        
-        assembleMyGroupDetail(container: container)
-        assembleMyGroupInfoEdit(container: container)
-        assembleMyGroupMemberEdit(container: container)
-        assembleMyGroupNoticeEdit(container: container)
-        
-        assembleMemberProfile(container: container)
-        
-        assembleSocialDailyCalendar(container: container)
-    }
-    
-}
-
-import Swinject
-
-extension PresentationAssembly {
-    func assembleMyPageMain(container: Container) {
-        container.register(MyPageMainViewModel.self) { (r, injectable: MyPageMainViewModel.Injectable) in
-            return MyPageMainViewModel(
-                useCases: .init(
-                    updateProfileUseCase: r.resolve(UpdateProfileUseCase.self)!,
-                    getTokenUseCase: r.resolve(GetTokenUseCase.self)!,
-                    refreshTokenUseCase: r.resolve(RefreshTokenUseCase.self)!,
-                    removeTokenUseCase: r.resolve(RemoveTokenUseCase.self)!,
-                    removeProfileUseCase: r.resolve(RemoveProfileUseCase.self)!,
-                    fetchImageUseCase: r.resolve(FetchImageUseCase.self)!,
-                    getSignedInSNSTypeUseCase: r.resolve(GetSignedInSNSTypeUseCase.self)!,
-                    convertToSha256UseCase: r.resolve(ConvertToSha256UseCase.self)!,
-                    revokeAppleTokenUseCase: r.resolve(RevokeAppleTokenUseCase.self)!
-                ),
-                injectable: injectable
-            )
-        }
-        
-        container.register(MyPageMainViewController.self) { (r, injectable: MyPageMainViewModel.Injectable) in
-            return MyPageMainViewController(viewModel: r.resolve(MyPageMainViewModel.self, argument: injectable)!)
-        }
-    }
-    
-    func assembleMyPageReadableViewModel(container: Container) {
-        container.register(MyPageReadableViewModel.self) { (r, injectable: MyPageReadableViewModel.Injectable) in
-            return MyPageReadableViewModel(
-                useCases: .init(),
-                injectable: injectable
-            )
-        }
-        
-        container.register(MyPageReadableViewController.self) { (r, injectable: MyPageReadableViewModel.Injectable) in
-            return MyPageReadableViewController(viewModel: r.resolve(MyPageReadableViewModel.self, argument: injectable)!)
-        }
-    }
-
-    
-    func assembleMyPageEnquire(container: Container) {
-        container.register(MyPageEnquireViewModel.self) { r in
-            return MyPageEnquireViewModel()
-        }
-        
-        container.register(MyPageEnquireViewController.self) { r in
-            return MyPageEnquireViewController(viewModel: r.resolve(MyPageEnquireViewModel.self)!)
-        }
-    }
-    
-    func assembleMyPageEdit(container: Container) {
-        container.register(MyPageEditViewModel.self) { (r, injectable: MyPageEditViewModel.Injectable) in
-            return MyPageEditViewModel(
-                useCases: .init(
-                    getTokenUseCase: r.resolve(GetTokenUseCase.self)!,
-                    refreshTokenUseCase: r.resolve(RefreshTokenUseCase.self)!,
-                    readProfileUseCase: r.resolve(ReadProfileUseCase.self)!,
-                    updateProfileUseCase: r.resolve(UpdateProfileUseCase.self)!,
-                    fetchImageUseCase: r.resolve(FetchImageUseCase.self)!
-                ),
-                injectable: injectable
-            )
-        }
-        
-        container.register(MyPageEditViewController.self) { (r, injectable: MyPageEditViewModel.Injectable) in
-            return MyPageEditViewController(viewModel: r.resolve(MyPageEditViewModel.self, argument: injectable)!)
-        }
-    }
-    
-}
-
-
-extension PresentationAssembly {
-    func assembleHomeCalendar(container: Container) {
-        container.register(HomeCalendarViewModel.self) { (r, injectable: HomeCalendarViewModel.Injectable) in
-            return HomeCalendarViewModel(
-                useCases: .init(
-                    getTokenUseCase: r.resolve(GetTokenUseCase.self)!,
-                    refreshTokenUseCase: r.resolve(RefreshTokenUseCase.self)!,
-                    createTodoUseCase: r.resolve(CreateTodoUseCase.self)!,
-                    readTodoListUseCase: r.resolve(ReadTodoListUseCase.self)!,
-                    updateTodoUseCase: r.resolve(UpdateTodoUseCase.self)!,
-                    deleteTodoUseCase: r.resolve(DeleteTodoUseCase.self)!,
-                    todoCompleteUseCase: r.resolve(TodoCompleteUseCase.self)!,
-                    createCategoryUseCase: r.resolve(CreateCategoryUseCase.self)!,
-                    readCategoryListUseCase: r.resolve(ReadCategoryListUseCase.self)!,
-                    updateCategoryUseCase: r.resolve(UpdateCategoryUseCase.self)!,
-                    deleteCategoryUseCase: r.resolve(DeleteCategoryUseCase.self)!,
-                    fetchGroupCategoryListUseCase: r.resolve(FetchAllGroupCategoryListUseCase.self)!,
-                    fetchMyGroupNameListUseCase: r.resolve(FetchMyGroupNameListUseCase.self)!,
-                    groupCreateUseCase: r.resolve(GroupCreateUseCase.self)!,
-                    withdrawGroupUseCase: r.resolve(WithdrawGroupUseCase.self)!,
-                    deleteGroupUseCase: r.resolve(DeleteGroupUseCase.self)!,
-                    createMonthlyCalendarUseCase: r.resolve(CreateMonthlyCalendarUseCase.self)!,
-                    dateFormatYYYYMMUseCase: r.resolve(DateFormatYYYYMMUseCase.self)!,
-                    readProfileUseCase: r.resolve(ReadProfileUseCase.self)!,
-                    updateProfileUseCase: r.resolve(UpdateProfileUseCase.self)!,
-                    fetchImageUseCase: r.resolve(FetchImageUseCase.self)!
-                ),
-                injectable: injectable
-            )
-        }
-        
-        container.register(HomeCalendarViewController.self) { (r, injectable: HomeCalendarViewModel.Injectable) in
-            return HomeCalendarViewController(viewModel: r.resolve(HomeCalendarViewModel.self, argument: injectable)!)
-        }
-    }
-}
-
-extension PresentationAssembly {
-    func assembleDailyCalendar(container: Container) {
-        container.register(DailyCalendarViewModel.self) { (r, injectable: DailyCalendarViewModel.Injectable) in
-            return DailyCalendarViewModel(
-                useCases: .init(
-                    getTokenUseCase: r.resolve(GetTokenUseCase.self)!,
-                    refreshTokenUseCase: r.resolve(RefreshTokenUseCase.self)!,
-                    createTodoUseCase: r.resolve(CreateTodoUseCase.self)!,
-                    updateTodoUseCase: r.resolve(UpdateTodoUseCase.self)!,
-                    deleteTodoUseCase: r.resolve(DeleteTodoUseCase.self)!,
-                    todoCompleteUseCase: r.resolve(TodoCompleteUseCase.self)!,
-                    createCategoryUseCase: r.resolve(CreateCategoryUseCase.self)!,
-                    updateCategoryUseCase: r.resolve(UpdateCategoryUseCase.self)!,
-                    deleteCategoryUseCase: r.resolve(DeleteCategoryUseCase.self)!,
-                    readCategoryUseCase: r.resolve(ReadCategoryListUseCase.self)!
-                ),
-                injectable: injectable
-            )
-        }
-        
-        container.register(DailyCalendarViewController.self) { (r, injectable: DailyCalendarViewModel.Injectable) in
-            return DailyCalendarViewController(viewModel: r.resolve(DailyCalendarViewModel.self, argument: injectable)!)
-        }
-    }
-}
-
-extension PresentationAssembly {
-    enum TodoDetailPageType: String {
-        case memberTodo = "MEMBER_TODO_DETAIL"
-        case socialTodo = "SOCIAL_TODO_DETAIL"
-    }
-    
-    func assembleTodoDetail(container: Container) {
-        container.register(
-            MemberTodoDetailViewModel.self
-        ) { (r, injectable: MemberTodoDetailViewModel.Injectable) in
-            return MemberTodoDetailViewModel(
-                getTokenUseCase: r.resolve(GetTokenUseCase.self)!,
-                refreshTokenUseCase: r.resolve(RefreshTokenUseCase.self)!,
-                createTodoUseCase: r.resolve(CreateTodoUseCase.self)!,
-                updateTodoUseCase: r.resolve(UpdateTodoUseCase.self)!,
-                deleteTodoUseCase: r.resolve(DeleteTodoUseCase.self)!,
-                createCategoryUseCase: r.resolve(CreateCategoryUseCase.self)!,
-                updateCategoryUseCase: r.resolve(UpdateCategoryUseCase.self)!,
-                deleteCategoryUseCase: r.resolve(DeleteCategoryUseCase.self)!,
-                readCategoryUseCase: r.resolve(ReadCategoryListUseCase.self)!,
-                injectable: injectable
-            )
-        }
-        
-        container.register(SocialTodoDetailViewModel.self) { (r, injectable: SocialTodoDetailViewModel.Injectable) in
-            return SocialTodoDetailViewModel(
-                getTokenUseCase: r.resolve(GetTokenUseCase.self)!,
-                refreshTokenUseCase: r.resolve(RefreshTokenUseCase.self)!,
-                fetchGroupMemberTodoDetailUseCase: r.resolve(FetchGroupMemberTodoDetailUseCase.self)!,
-                fetchGroupTodoDetailUseCase: r.resolve(FetchGroupTodoDetailUseCase.self)!,
-                createGroupTodoUseCase: r.resolve(CreateGroupTodoUseCase.self)!,
-                updateGroupTodoUseCase: r.resolve(UpdateGroupTodoUseCase.self)!,
-                deleteGroupTodoUseCase: r.resolve(DeleteGroupTodoUseCase.self)!,
-                createGroupCategoryUseCase: r.resolve(CreateGroupCategoryUseCase.self)!,
-                updateGroupCategoryUseCase: r.resolve(UpdateGroupCategoryUseCase.self)!,
-                deleteGroupCategoryUseCase: r.resolve(DeleteGroupCategoryUseCase.self)!,
-                fetchGroupCategorysUseCase: r.resolve(FetchGroupCategorysUseCase.self)!,
-                injectable: injectable
-            )
-        }
-        
-        container.register(
-            TodoDetailViewController.self,
-            name: TodoDetailPageType.memberTodo.rawValue
-        ) { (r, injectable: MemberTodoDetailViewModel.Injectable) in
-            return TodoDetailViewController(viewModel: r.resolve(MemberTodoDetailViewModel.self, argument: injectable)!)
-        }
-        
-        container.register(TodoDetailViewController.self, name: TodoDetailPageType.socialTodo.rawValue) { (r, injectable: SocialTodoDetailViewModel.Injectable) in
-            return TodoDetailViewController(viewModel: r.resolve(SocialTodoDetailViewModel.self, argument: injectable)!)
-        }
-    }
-}
-
-
-extension PresentationAssembly {
-    func assembleSignIn(container: Container) {
-        container.register(SignInViewModel.self) { (r, injectable: SignInViewModel.Injectable) in
-            return SignInViewModel(
-                useCases: SignInViewModel.UseCases(
-                    kakaoSignInUseCase: r.resolve(KakaoSignInUseCase.self)!,
-                    googleSignInUseCase: r.resolve(GoogleSignInUseCase.self)!,
-                    appleSignInUseCase: r.resolve(AppleSignInUseCase.self)!,
-                    convertToSha256UseCase: r.resolve(ConvertToSha256UseCase.self)!,
-                    setSignedInSNSTypeUseCase: r.resolve(SetSignedInSNSTypeUseCase.self)!,
-                    revokeAppleTokenUseCase: r.resolve(RevokeAppleTokenUseCase.self)!,
-                    setTokenUseCase: r.resolve(SetTokenUseCase.self)!
-                ),
-                injectable: injectable
-            )
-        }
-        
-        container.register(SignInViewController.self) { (r, injectable: SignInViewModel.Injectable) in
-            return SignInViewController(viewModel: r.resolve(SignInViewModel.self, argument: injectable)!)
-        }
-    }
-    
-    func assembleRedirectionalWeb(container: Container) {
-        container.register(RedirectionalWebViewModel.self) { (r, injectable: RedirectionalWebViewModel.Injectable) in
-            return RedirectionalWebViewModel(useCases: .init(), injectable: injectable)
-        }
-        
-        container.register(RedirectionalWebViewController.self) { (r, injectable: RedirectionalWebViewModel.Injectable) in
-            return RedirectionalWebViewController(viewModel: r.resolve(RedirectionalWebViewModel.self, argument: injectable)!)
-        }
-    }
-}
-
-extension PresentationAssembly {
-    func assembleSearchHome(container: Container) {
-        container.register(SearchHomeViewModel.self) { (r, injectable: SearchHomeViewModel.Injectable) in
-            return SearchHomeViewModel(
-                useCases: .init(
-                    getTokenUseCase: r.resolve(GetTokenUseCase.self)!,
-                    refreshTokenUseCase: r.resolve(RefreshTokenUseCase.self)!,
-                    fetchSearchHomeUseCase: r.resolve(FetchSearchHomeUseCase.self)!,
-                    fetchImageUseCase: r.resolve(FetchImageUseCase.self)!
-                ),
-                injectable: injectable
-            )
-        }
-        
-        container.register(SearchHomeViewController.self) { (r, injectable: SearchHomeViewModel.Injectable) in
-            return SearchHomeViewController(viewModel: r.resolve(SearchHomeViewModel.self, argument: injectable)!)
-        }
-    }
-    
-    func assembleSearchResult(container: Container) {
-        container.register(SearchResultViewModel.self) { (r, injectable: SearchResultViewModel.Injectable) in
-            return SearchResultViewModel(
-                useCases: .init(
-                    recentQueryRepository: r.resolve(RecentQueryRepository.self)!,
-                    getTokenUseCase: r.resolve(GetTokenUseCase.self)!,
-                    refreshTokenUseCase: r.resolve(RefreshTokenUseCase.self)!,
-                    fetchSearchResultUseCase: r.resolve(FetchSearchResultUseCase.self)!,
-                    fetchImageUseCase: r.resolve(FetchImageUseCase.self)!
-                ),
-                injectable: injectable
-            )
-        }
-        
-        container.register(SearchHomeViewController.self) { (r, injectable: SearchHomeViewModel.Injectable) in
-            return SearchHomeViewController(viewModel: r.resolve(SearchHomeViewModel.self, argument: injectable)!)
-        }
-    }
-}
-
-extension PresentationAssembly {
-    func assembleGroupCreate(container: Container) {
-        container.register(GroupCreateViewModel.self) { (r, injectable: GroupCreateViewModel.Injectable) in
-            return GroupCreateViewModel(
-                useCases: .init(
-                    getTokenUseCase: r.resolve(GetTokenUseCase.self)!,
-                    refreshTokenUseCase: r.resolve(RefreshTokenUseCase.self)!,
-                    groupCreateUseCase: r.resolve(GroupCreateUseCase.self)!
-                ),
-                injectable: injectable
-            )
-        }
-        
-        container.register(GroupCreateViewController.self) { (r, injectable: GroupCreateViewModel.Injectable) in
-            return GroupCreateViewController(viewModel: r.resolve(GroupCreateViewModel.self, argument: injectable)!)
-        }
-    }
-    
-    func assembleGroupCreateLoading(container: Container) {
-        container.register(GroupCreateLoadViewModel.self) { (r, injectable: GroupCreateLoadViewModel.Injectable) in
-            return GroupCreateLoadViewModel(
-                useCases: .init(
-                    getTokenUseCase: r.resolve(GetTokenUseCase.self)!,
-                    refreshTokenUseCase: r.resolve(RefreshTokenUseCase.self)!,
-                    groupCreateUseCase: r.resolve(GroupCreateUseCase.self)!
-                ),
-                injectable: injectable
-            )
-        }
-        
-        container.register(GroupCreateLoadViewController.self) { (r, injectable: GroupCreateLoadViewModel.Injectable) in
-            return GroupCreateLoadViewController(viewModel: r.resolve(GroupCreateLoadViewModel.self, argument: injectable)!)
-        }
-    }
-}
-
-extension PresentationAssembly {
-    func assembleGroupIntroduce(container: Container) {
-        container.register(GroupIntroduceViewModel.self) { (r, injectable: GroupIntroduceViewModel.Injectable) in
-            return GroupIntroduceViewModel(
-                useCases: .init(
-                    getTokenUseCase: r.resolve(GetTokenUseCase.self)!,
-                    refreshTokenUseCase: r.resolve(RefreshTokenUseCase.self)!,
-                    setTokenUseCase: r.resolve(SetTokenUseCase.self)!,
-                    fetchUnJoinedGroupUseCase: r.resolve(FetchUnJoinedGroupUseCase.self)!,
-                    fetchMemberListUseCase: r.resolve(FetchMemberListUseCase.self)!,
-                    fetchImageUseCase: r.resolve(FetchImageUseCase.self)!,
-                    applyGroupJoinUseCase: r.resolve(ApplyGroupJoinUseCase.self)!,
-                    generateGroupLinkUseCase: r.resolve(GenerateGroupLinkUseCase.self)!
-                ),
-                injectable: injectable
-            )
-        }
-        
-        container.register(GroupIntroduceViewController.self) { (r, injectable: GroupIntroduceViewModel.Injectable) in
-            return GroupIntroduceViewController(viewModel: r.resolve(GroupIntroduceViewModel.self, argument: injectable)!)
-        }
-    }
-}
-
-extension PresentationAssembly {
-    func assembleNotification(container: Container) {
-        container.register(NotificationViewModel.self) { (r, injectable: NotificationViewModel.Injectable) in
-            return NotificationViewModel(
-                useCases: .init(
-                    getTokenUseCase: r.resolve(GetTokenUseCase.self)!,
-                    refreshTokenUseCase: r.resolve(RefreshTokenUseCase.self)!,
-                    setTokenUseCase: r.resolve(SetTokenUseCase.self)!,
-                    fetchJoinApplyListUseCase: r.resolve(FetchJoinApplyListUseCase.self)!,
-                    fetchImageUseCase: r.resolve(FetchImageUseCase.self)!,
-                    acceptGroupJoinUseCase: r.resolve(AcceptGroupJoinUseCase.self)!,
-                    denyGroupJoinUseCase: r.resolve(DenyGroupJoinUseCase.self)!
-                ),
-                injectable: injectable
-            )
-        }
-        
-        container.register(NotificationViewController.self) { (r, injectable: NotificationViewModel.Injectable) in
-            return NotificationViewController(viewModel: r.resolve(NotificationViewModel.self, argument: injectable)!)
-        }
-    }
-}
-
-extension PresentationAssembly {
-    func assembleMyGroupList(container: Container) {
-        container.register(MyGroupListViewModel.self) { (r, injectable: MyGroupListViewModel.Injectable) in
-            return MyGroupListViewModel(
-                useCases: .init(
-                    getTokenUseCase: r.resolve(GetTokenUseCase.self)!,
-                    refreshTokenUsecase: r.resolve(RefreshTokenUseCase.self)!,
-                    setTokenUseCase: r.resolve(SetTokenUseCase.self)!,
-                    fetchMyGroupListUseCase: r.resolve(FetchMyGroupListUseCase.self)!,
-                    fetchImageUseCase: r.resolve(FetchImageUseCase.self)!,
-                    groupCreateUseCase: r.resolve(GroupCreateUseCase.self)!,
-                    setOnlineUseCase: r.resolve(SetOnlineUseCase.self)!,
-                    updateGroupInfoUseCase: r.resolve(UpdateGroupInfoUseCase.self)!,
-                    withdrawGroupUseCase: r.resolve(WithdrawGroupUseCase.self)!,
-                    deleteGroupUseCase: r.resolve(DeleteGroupUseCase.self)!
-                ),
-                injectable: injectable
-            )
-        }
-        
-        container.register(MyGroupListViewController.self) { (r, injectable: MyGroupListViewModel.Injectable) in
-            return MyGroupListViewController(viewModel: r.resolve(MyGroupListViewModel.self, argument: injectable)!)
-        }
-    }
-}
-
-extension PresentationAssembly {
-    func assembleMyGroupDetail(container: Container) {
-        container.register(MyGroupDetailViewModel.self) { (r, injectable: MyGroupDetailViewModel.Injectable) in
-            return MyGroupDetailViewModel(
-                useCases: .init(
-                    fetchMyGroupDetailUseCase: r.resolve(FetchMyGroupDetailUseCase.self)!,
-                    updateNoticeUseCase: r.resolve(UpdateNoticeUseCase.self)!,
-                    updateInfoUseCase: r.resolve(UpdateGroupInfoUseCase.self)!,
-                    withdrawGroupUseCase: r.resolve(WithdrawGroupUseCase.self)!,
-                    fetchMyGroupMemberListUseCase: r.resolve(FetchMyGroupMemberListUseCase.self)!,
-                    fetchImageUseCase: r.resolve(FetchImageUseCase.self)!,
-                    memberKickOutUseCase: r.resolve(MemberKickOutUseCase.self)!,
-                    setOnlineUseCase: r.resolve(SetOnlineUseCase.self)!,
-                    getTokenUseCase: r.resolve(GetTokenUseCase.self)!,
-                    refreshTokenUseCase: r.resolve(RefreshTokenUseCase.self)!,
-                    createMonthlyCalendarUseCase: r.resolve(CreateMonthlyCalendarUseCase.self)!,
-                    fetchMyGroupCalendarUseCase: r.resolve(FetchGroupMonthlyCalendarUseCase.self)!,
-                    createGroupTodoUseCase: r.resolve(CreateGroupTodoUseCase.self)!,
-                    updateGroupTodoUseCase: r.resolve(UpdateGroupTodoUseCase.self)!,
-                    deleteGroupTodoUseCase: r.resolve(DeleteGroupTodoUseCase.self)!,
-                    updateGroupCategoryUseCase: r.resolve(UpdateGroupCategoryUseCase.self)!,
-                    generateGroupLinkUseCase: r.resolve(GenerateGroupLinkUseCase.self)!
-                ),
-                injectable: injectable
-            )
-        }
-        
-        container.register(MyGroupDetailViewController.self) { (r, injectable: MyGroupDetailViewModel.Injectable) in
-            return MyGroupDetailViewController(viewModel: r.resolve(MyGroupDetailViewModel.self, argument: injectable)!)
-        }
-    }
-    
-    func assembleMyGroupInfoEdit(container: Container) {
-        container.register(MyGroupInfoEditViewModel.self) { (r, injectable: MyGroupInfoEditViewModel.Injectable) in
-            return MyGroupInfoEditViewModel(
-                useCases: .init(
-                    getTokenUseCase: r.resolve(GetTokenUseCase.self)!,
-                    refreshTokenUseCase: r.resolve(RefreshTokenUseCase.self)!,
-                    fetchImageUseCase: r.resolve(FetchImageUseCase.self)!,
-                    updateGroupInfoUseCase: r.resolve(UpdateGroupInfoUseCase.self)!,
-                    deleteGroupUseCase: r.resolve(DeleteGroupUseCase.self)!
-                ),
-                injectable: injectable
-            )
-        }
-        
-        container.register(MyGroupInfoEditViewController.self) { (r, injectable: MyGroupInfoEditViewModel.Injectable) in
-            return MyGroupInfoEditViewController(viewModel: r.resolve(MyGroupInfoEditViewModel.self, argument: injectable)!)
-        }
-    }
-    
-    func assembleMyGroupMemberEdit(container: Container) {
-        container.register(MyGroupMemberEditViewModel.self) { (r, injectable: MyGroupMemberEditViewModel.Injectable) in
-            return MyGroupMemberEditViewModel(
-                useCase: .init(
-                    getTokenUseCase: r.resolve(GetTokenUseCase.self)!,
-                    refreshTokenUseCase: r.resolve(RefreshTokenUseCase.self)!,
-                    fetchMyGroupMemberListUseCase: r.resolve(FetchMyGroupMemberListUseCase.self)!,
-                    fetchImageUseCase: r.resolve(FetchImageUseCase.self)!,
-                    memberKickOutUseCase: r.resolve(MemberKickOutUseCase.self)!
-                ),
-                injectable: injectable
-            )
-        }
-        
-        container.register(MyGroupMemberEditViewController.self) { (r, injectable: MyGroupMemberEditViewModel.Injectable) in
-            return MyGroupMemberEditViewController(viewModel: r.resolve(MyGroupMemberEditViewModel.self, argument: injectable)!)
-        }
-    }
-    
-    func assembleMyGroupNoticeEdit(container: Container) {
-        container.register(MyGroupNoticeEditViewModel.self) { (r, injectable: MyGroupNoticeEditViewModel.Injectable) in
-            return MyGroupNoticeEditViewModel(
-                useCases: .init(
-                    getTokenUseCase: r.resolve(GetTokenUseCase.self)!,
-                    refreshTokenUseCase: r.resolve(RefreshTokenUseCase.self)!,
-                    updateNoticeUseCase: r.resolve(UpdateNoticeUseCase.self)!
-                ),
-                injectable: injectable
-            )
-        }
-        
-        container.register(MyGroupNoticeEditViewController.self) { (r, injectable: MyGroupNoticeEditViewModel.Injectable) in
-            return MyGroupNoticeEditViewController(viewModel: r.resolve(MyGroupNoticeEditViewModel.self, argument: injectable)!)
-        }
-    }
-}
-
-extension PresentationAssembly {
-    func assembleMemberProfile(container: Container) {
-        container.register(MemberProfileViewModel.self) { (r, injectable: MemberProfileViewModel.Injectable) in
-            return MemberProfileViewModel(
-                useCases: .init(
-                    createMonthlyCalendarUseCase: r.resolve(CreateMonthlyCalendarUseCase.self)!,
-                    dateFormatYYYYMMUseCase: r.resolve(DateFormatYYYYMMUseCase.self)!,
-                    getTokenUseCase: r.resolve(GetTokenUseCase.self)!,
-                    refreshTokenUseCase: r.resolve(RefreshTokenUseCase.self)!,
-                    fetchMemberCalendarUseCase: r.resolve(FetchGroupMemberCalendarUseCase.self)!,
-                    fetchImageUseCase: r.resolve(FetchImageUseCase.self)!
-                ),
-                injectable: injectable
-            )
-        }
-        
-        container.register(MemberProfileViewController.self) { (r, injectable: MemberProfileViewModel.Injectable) in
-            return MemberProfileViewController(viewModel: r.resolve(MemberProfileViewModel.self, argument: injectable)!)
-        }
-    }
-}
-
-extension PresentationAssembly {
-    func assembleSocialDailyCalendar(container: Container) {
-        container.register(SocialDailyCalendarViewModel.self) { (r, injectable: SocialDailyCalendarViewModel.Injectable) in
-            return SocialDailyCalendarViewModel(
-                useCases: .init(
-                    getTokenUseCase: r.resolve(GetTokenUseCase.self)!,
-                    refreshTokenUseCase: r.resolve(RefreshTokenUseCase.self)!,
-                    fetchGroupDailyTodoListUseCase: r.resolve(FetchGroupDailyCalendarUseCase.self)!,
-                    fetchMemberDailyCalendarUseCase: r.resolve(FetchGroupMemberDailyCalendarUseCase.self)!,
-                    createGroupTodoUseCase: r.resolve(CreateGroupTodoUseCase.self)!,
-                    updateGroupTodoUseCase: r.resolve(UpdateGroupTodoUseCase.self)!,
-                    deleteGroupTodoUseCase: r.resolve(DeleteGroupTodoUseCase.self)!,
-                    updateGroupCategoryUseCase: r.resolve(UpdateGroupCategoryUseCase.self)!
-                ),
-                injectable: injectable
-            )
-        }
-        
-        container.register(SocialDailyCalendarViewController.self) { (r, injectable: SocialDailyCalendarViewModel.Injectable) in
-            return SocialDailyCalendarViewController(viewModel: r.resolve(SocialDailyCalendarViewModel.self, argument: injectable)!)
-        }
-    }
-}
